+++ pygpsclient/serial_handler.py
@@ -86,7 +86,7 @@ class SerialHandler:
                 timeout=serial_settings.timeout,
             )
             self._serial_buffer = BufferedReader(self._serial_object)
-            self.__app.frm_banner.update_conn_status(CONNECTED)
+            self.__app.conn_status = CONNECTED
             self.__app.set_connection(
                 (
                     f"{serial_settings.port}:{serial_settings.port_desc} "
@@ -94,19 +94,13 @@ class SerialHandler:
                 ),
                 "green",
             )
-            self.__app.frm_settings.enable_controls(CONNECTED)
             self._connected = True
             self.start_read_thread()
-
-            if self.__app.frm_settings.datalogging:
-                self.__app.file_handler.open_logfile()
-
-            if self.__app.frm_settings.record_track:
-                self.__app.file_handler.open_trackfile()
             self.__app.set_status("Connected", "blue")
 
         except (IOError, SerialException, SerialTimeoutException) as err:
             self._connected = False
+            self.__app.conn_status = DISCONNECTED
             self.__app.set_connection(
                 (
                     f"{serial_settings.port}:{serial_settings.port_desc} "
@@ -115,8 +109,6 @@ class SerialHandler:
                 "red",
             )
             self.__app.set_status(SEROPENERROR.format(err), "red")
-            self.__app.frm_banner.update_conn_status(DISCONNECTED)
-            self.__app.frm_settings.enable_controls(DISCONNECTED)
 
     def connect_file(self):
         """
@@ -131,24 +123,15 @@ class SerialHandler:
         try:
             self._serial_object = open(in_filepath, "rb")
             self._serial_buffer = BufferedReader(self._serial_object)
-            self.__app.frm_banner.update_conn_status(CONNECTED_FILE)
+            self.__app.conn_status = CONNECTED_FILE
             self.__app.set_connection(f"{in_filepath}", "blue")
-            self.__app.frm_settings.enable_controls(CONNECTED_FILE)
             self._connectedfile = True
             self.start_readfile_thread()
-
-            if self.__app.frm_settings.datalogging:
-                self.__app.file_handler.open_logfile()
-
-            if self.__app.frm_settings.record_track:
-                self.__app.file_handler.open_trackfile()
-
         except (IOError, SerialException, SerialTimeoutException) as err:
             self._connected = False
+            self.__app.conn_status = DISCONNECTED
             self.__app.set_connection(f"{in_filepath}", "red")
             self.__app.set_status(SEROPENERROR.format(err), "red")
-            self.__app.frm_banner.update_conn_status(DISCONNECTED)
-            self.__app.frm_settings.enable_controls(DISCONNECTED)
 
     def disconnect(self):
         """
@@ -159,16 +142,7 @@ class SerialHandler:
             try:
                 self._reading = False
                 self._serial_object.close()
-                self.__app.frm_banner.update_conn_status(DISCONNECTED)
-                self.__app.set_connection(NOTCONN, "red")
-                self.__app.set_status("", "blue")
-
-                if self.__app.frm_settings.datalogging:
-                    self.__app.file_handler.close_logfile()
-
-                if self.__app.frm_settings.record_track:
-                    self.__app.file_handler.close_trackfile()
-
+                self.__app.conn_status = DISCONNECTED
             except (SerialException, SerialTimeoutException):
                 pass