+++ pygpsclient/settings_frame.py
@@ -34,6 +34,7 @@ from tkinter import (
 from PIL import ImageTk, Image
 import pyubx2.ubxtypes_core as ubt
 from pygpsclient.serialconfig_frame import SerialConfigFrame
+from pygpsclient.socketconfig_frame import SocketConfigFrame
 from pygpsclient.globals import (
     ENTCOL,
     DDD,
@@ -45,10 +46,13 @@ from pygpsclient.globals import (
     UIK,
     READONLY,
     CONNECTED,
+    CONNECTED_SOCKET,
     CONNECTED_FILE,
     DISCONNECTED,
     NOPORTS,
     ICON_CONN,
+    ICON_SERIAL,
+    ICON_SOCKET,
     ICON_DISCONN,
     ICON_UBXCONFIG,
     ICON_NTRIPCONFIG,
@@ -121,6 +125,8 @@ class SettingsFrame(Frame):
         self._trackpath = None
         self._display_format = StringVar()
         self._img_conn = ImageTk.PhotoImage(Image.open(ICON_CONN))
+        self._img_serial = ImageTk.PhotoImage(Image.open(ICON_SERIAL))
+        self._img_socket = ImageTk.PhotoImage(Image.open(ICON_SOCKET))
         self._img_disconn = ImageTk.PhotoImage(Image.open(ICON_DISCONN))
         self._img_ubxconfig = ImageTk.PhotoImage(Image.open(ICON_UBXCONFIG))
         self._img_ntripconfig = ImageTk.PhotoImage(Image.open(ICON_NTRIPCONFIG))
@@ -146,23 +152,27 @@ class SettingsFrame(Frame):
             self, preselect=KNOWNGPS, timeouts=TIMEOUTS, bpsrates=BPSRATES
         )
 
+        # socket configuration panel
+        self._frm_socket = SocketConfigFrame(self)
+
         # connection buttons
         self._frm_buttons = Frame(self)
         self._btn_connect = Button(
             self._frm_buttons,
             width=45,
             height=35,
-            image=self._img_conn,
+            image=self._img_serial,
             command=lambda: self.__app.serial_handler.connect(),
         )
-        self._btn_disconnect = Button(
+        self._lbl_connect = Label(self._frm_buttons, text="USB/UART")
+        self._btn_connect_socket = Button(
             self._frm_buttons,
             width=45,
             height=35,
-            image=self._img_disconn,
-            command=lambda: self.__app.serial_handler.disconnect(),
-            state=DISABLED,
+            image=self._img_socket,
+            command=lambda: self.__app.socket_handler.connect(),
         )
+        self._lbl_connect_socket = Label(self._frm_buttons, text="TCP/UDP")
         self._btn_connect_file = Button(
             self._frm_buttons,
             width=45,
@@ -170,6 +180,16 @@ class SettingsFrame(Frame):
             image=self._img_dataread,
             command=lambda: self._on_data_stream(),
         )
+        self._lbl_connect_file = Label(self._frm_buttons, text="FILE")
+        self._btn_disconnect = Button(
+            self._frm_buttons,
+            width=45,
+            height=35,
+            image=self._img_disconn,
+            command=lambda: self._disconnect(),  # self.__app.serial_handler.disconnect(),
+            state=DISABLED,
+        )
+        self._lbl_disconnect = Label(self._frm_buttons, text="STOP")
         self._lbl_status_preset = Label(
             self._frm_buttons, font=self.__app.font_md2, text=""
         )
@@ -236,10 +256,11 @@ class SettingsFrame(Frame):
         )
         self._chk_webmap = Checkbutton(
             self._frm_options,
-            text="Web Map    Zoom",
+            text="Web Map",
             variable=self._webmap,
             command=lambda: self._on_webmap(),
         )
+        self._lbl_mapzoom = Label(self._frm_options, text="Zoom")
         self._scl_mapzoom = Scale(
             self._frm_options,
             from_=1,
@@ -303,62 +324,82 @@ class SettingsFrame(Frame):
         """
 
         self._frm_serial.grid(
-            column=0, row=1, columnspan=4, padx=3, pady=3, sticky=(W, E)
+            column=0, row=1, columnspan=4, padx=2, pady=2, sticky=(W, E)
         )
         ttk.Separator(self).grid(
-            column=0, row=2, columnspan=4, padx=3, pady=3, sticky=(W, E)
+            column=0, row=2, columnspan=4, padx=2, pady=2, sticky=(W, E)
         )
 
-        self._frm_buttons.grid(column=0, row=3, columnspan=4, sticky=(W, E))
-        self._btn_connect.grid(column=0, row=0, padx=3, pady=3)
-        self._btn_connect_file.grid(column=1, row=0, padx=3, pady=3)
-        self._btn_disconnect.grid(column=3, row=0, padx=3, pady=3)
+        self._frm_socket.grid(
+            column=0, row=3, columnspan=4, padx=2, pady=2, sticky=(W, E)
+        )
+        ttk.Separator(self).grid(
+            column=0, row=4, columnspan=4, padx=2, pady=2, sticky=(W, E)
+        )
+
+        self._frm_buttons.grid(column=0, row=5, columnspan=4, sticky=(W, E))
+        self._btn_connect.grid(column=0, row=0, padx=2, pady=1)
+        self._btn_connect_socket.grid(column=1, row=0, padx=2, pady=1)
+        self._btn_connect_file.grid(column=2, row=0, padx=2, pady=1)
+        self._btn_disconnect.grid(column=3, row=0, padx=2, pady=1)
+        self._lbl_connect.grid(column=0, row=1, padx=1, pady=1, sticky=(W, E))
+        self._lbl_connect_socket.grid(column=1, row=1, padx=1, pady=1, sticky=(W, E))
+        self._lbl_connect_file.grid(column=2, row=1, padx=1, pady=1, sticky=(W, E))
+        self._lbl_disconnect.grid(column=3, row=1, padx=1, pady=1, sticky=(W, E))
 
         ttk.Separator(self).grid(
-            column=0, row=7, columnspan=4, padx=3, pady=3, sticky=(W, E)
+            column=0, row=7, columnspan=4, padx=2, pady=2, sticky=(W, E)
         )
 
         self._frm_options.grid(column=0, row=8, columnspan=4, sticky=(W, E))
-        self._lbl_protocol.grid(column=0, row=0, padx=3, pady=3, sticky=(W))
+        self._lbl_protocol.grid(column=0, row=0, padx=2, pady=2, sticky=(W))
         self._chk_nmea.grid(column=1, row=0, padx=0, pady=0, sticky=(W))
         self._chk_ubx.grid(column=2, row=0, padx=0, pady=0, sticky=(W))
         self._chk_rtcm.grid(column=3, row=0, padx=0, pady=0, sticky=(W))
-        self._lbl_consoledisplay.grid(column=0, row=1, padx=2, pady=3, sticky=(W))
+        self._lbl_consoledisplay.grid(column=0, row=1, padx=2, pady=2, sticky=(W))
         self._spn_conformat.grid(
-            column=1, row=1, columnspan=2, padx=1, pady=3, sticky=(W)
+            column=1, row=1, columnspan=2, padx=1, pady=2, sticky=(W)
         )
-        self._lbl_format.grid(column=0, row=2, padx=3, pady=3, sticky=(W))
-        self._spn_format.grid(column=1, row=2, padx=2, pady=3, sticky=(W))
-        self._lbl_units.grid(column=0, row=3, padx=3, pady=3, sticky=(W))
-        self._spn_units.grid(column=1, row=3, columnspan=3, padx=2, pady=3, sticky=(W))
-        self._chk_scroll.grid(column=0, row=4, padx=3, pady=3, sticky=(W))
+        self._lbl_format.grid(column=0, row=2, padx=2, pady=2, sticky=(W))
+        self._spn_format.grid(column=1, row=2, padx=2, pady=2, sticky=(W))
+        self._lbl_units.grid(column=0, row=3, padx=2, pady=2, sticky=(W))
+        self._spn_units.grid(column=1, row=3, columnspan=3, padx=2, pady=2, sticky=(W))
+        self._chk_scroll.grid(column=0, row=4, padx=2, pady=2, sticky=(W))
         self._spn_maxlines.grid(
-            column=1, row=4, columnspan=3, padx=3, pady=3, sticky=(W)
+            column=1, row=4, columnspan=3, padx=2, pady=2, sticky=(W)
         )
-        self._chk_webmap.grid(column=0, row=5, padx=3, pady=3, sticky=(W))
-        self._scl_mapzoom.grid(column=1, row=5, columnspan=3, sticky=(W))
-        self._chk_legend.grid(column=0, row=6, padx=3, pady=3, sticky=(W))
+        self._chk_webmap.grid(column=0, row=5, padx=2, pady=2, sticky=(W))
+        self._lbl_mapzoom.grid(column=1, row=5, sticky=(E))
+        self._scl_mapzoom.grid(column=2, row=5, columnspan=2, sticky=(W))
+        self._chk_legend.grid(column=0, row=6, padx=2, pady=2, sticky=(W))
         self._chk_unusedsat.grid(
-            column=1, row=6, columnspan=3, padx=3, pady=3, sticky=(W)
+            column=1, row=6, columnspan=3, padx=2, pady=2, sticky=(W)
         )
-        self._chk_datalog.grid(column=0, row=7, padx=3, pady=3, sticky=(W))
+        self._chk_datalog.grid(column=0, row=7, padx=2, pady=2, sticky=(W))
         self._spn_datalog.grid(
-            column=1, row=7, columnspan=2, padx=3, pady=3, sticky=(W)
+            column=1, row=7, columnspan=2, padx=2, pady=2, sticky=(W)
         )
         self._chk_recordtrack.grid(
-            column=0, row=8, columnspan=2, padx=3, pady=3, sticky=(W)
+            column=0, row=8, columnspan=2, padx=2, pady=2, sticky=(W)
         )
 
         ttk.Separator(self._frm_options).grid(
-            column=0, row=9, columnspan=4, padx=3, pady=3, sticky=(W, E)
+            column=0, row=9, columnspan=4, padx=2, pady=2, sticky=(W, E)
         )
-        self._lbl_ubxconfig.grid(column=0, row=10, padx=3, pady=3, sticky=(W))
-        self._btn_ubxconfig.grid(column=1, row=10, padx=3, pady=3, sticky=(W))
-        ttk.Separator(self._frm_options).grid(
-            column=0, row=11, columnspan=4, padx=3, pady=3, sticky=(W, E)
-        )
-        self._lbl_ntripconfig.grid(column=0, row=12, padx=3, pady=3, sticky=(W))
-        self._btn_ntripconfig.grid(column=1, row=12, padx=3, pady=3, sticky=(W))
+        self._lbl_ubxconfig.grid(column=0, row=10, padx=2, pady=2, sticky=(E))
+        self._btn_ubxconfig.grid(column=1, row=10, padx=2, pady=2, sticky=(W))
+        self._lbl_ntripconfig.grid(column=2, row=10, padx=2, pady=2, sticky=(E))
+        self._btn_ntripconfig.grid(column=3, row=10, padx=2, pady=2, sticky=(W))
+
+    def _disconnect(self):
+
+        status = self.__app.conn_status
+        if status == CONNECTED_SOCKET:
+            self.__app.socket_handler.disconnect()
+        elif status == CONNECTED_FILE:
+            self.__app.serial_handler.disconnect()
+        elif status == CONNECTED:
+            self.__app.serial_handler.disconnect()
 
     def _on_ubx_config(self, *args, **kwargs):  # pylint: disable=unused-argument
         """
@@ -390,12 +431,13 @@ class SettingsFrame(Frame):
             self._logpath = self.__app.file_handler.set_logfile_path()
             if self._logpath is not None:
                 self.__app.set_status("Data logging enabled: " + self._logpath, "green")
+                self.__app.file_handler.open_logfile()
             else:
                 self._datalog.set(False)
         else:
             self._logpath = None
             self._datalog.set(False)
-            #             self.__app.file_handler.close_logfile()
+            self.__app.file_handler.close_logfile()
             self.__app.set_status("Data logging disabled", "blue")
 
     def _on_record_track(self):
@@ -409,12 +451,13 @@ class SettingsFrame(Frame):
                 self.__app.set_status(
                     "Track recording enabled: " + self._trackpath, "green"
                 )
+                self.__app.file_handler.open_trackfile()
             else:
                 self._record_track.set(False)
         else:
             self._trackpath = None
             self._record_track.set(False)
-            #             self.__app.file_handler.close_trackfile()
+            self.__app.file_handler.close_trackfile()
             self.__app.set_status("Track recording disabled", "blue")
 
     def _on_data_stream(self):
@@ -460,30 +503,52 @@ class SettingsFrame(Frame):
         """
 
         self._frm_serial.set_status(status)
+        self._frm_socket.set_status(status)
 
         self._btn_connect.config(
             state=(
-                DISABLED if status in (CONNECTED, CONNECTED_FILE, NOPORTS) else NORMAL
+                DISABLED
+                if status in (CONNECTED, CONNECTED_SOCKET, CONNECTED_FILE, NOPORTS)
+                else NORMAL
+            )
+        )
+        self._btn_connect_socket.config(
+            state=(
+                DISABLED
+                if status in (CONNECTED, CONNECTED_SOCKET, CONNECTED_FILE)
+                else NORMAL
             )
         )
         self._btn_disconnect.config(
-            state=(DISABLED if status in (DISCONNECTED, NOPORTS) else NORMAL)
+            state=(DISABLED if status in (DISCONNECTED,) else NORMAL)
         )
         self._chk_datalog.config(
             state=(
-                DISABLED if status in (CONNECTED, CONNECTED_FILE, NOPORTS) else NORMAL
+                DISABLED
+                if status in (CONNECTED, CONNECTED_SOCKET, CONNECTED_FILE)
+                else NORMAL
             )
         )
         self._spn_datalog.config(
             state=(
-                DISABLED if status in (CONNECTED, CONNECTED_FILE, NOPORTS) else READONLY
+                DISABLED
+                if status in (CONNECTED, CONNECTED_SOCKET, CONNECTED_FILE)
+                else READONLY
             )
         )
         self._chk_recordtrack.config(
-            state=(DISABLED if status in (CONNECTED, CONNECTED_FILE) else NORMAL)
+            state=(
+                DISABLED
+                if status in (CONNECTED, CONNECTED_SOCKET, CONNECTED_FILE)
+                else NORMAL
+            )
         )
         self._btn_connect_file.config(
-            state=(DISABLED if status in (CONNECTED, CONNECTED_FILE) else NORMAL)
+            state=(
+                DISABLED
+                if status in (CONNECTED, CONNECTED_SOCKET, CONNECTED_FILE)
+                else NORMAL
+            )
         )
 
     def get_size(self) -> tuple:
@@ -507,6 +572,16 @@ class SettingsFrame(Frame):
         """
 
         return self._frm_serial
+
+    def socket_settings(self) -> Frame:
+        """
+        Return reference to common socket configuration panel
+
+        :return: reference to socket form
+        :rtype: Frame
+        """
+
+        return self._frm_socket
 
     @property
     def protocol(self) -> int: